refactor: rename `cmd.py` to `command.py`
